# backend/tests/contracts/openapi.corvus-locomo.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Agentic institutional memory for coding harnesses. Situated episodes with outcomes, evidence-gated lessons, temporal supersession.",
    "title": "Corvus LoCoMo Eval",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auditor/audit/{neuron_id}": {
      "post": {
        "description": "Event-driven pass: reopen one memory now (contradicted attribution,\nuser correction, incident participation).",
        "operationId": "audit_one_auditor_audit__neuron_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "neuron_id",
            "required": true,
            "schema": {
              "title": "Neuron Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit One",
        "tags": [
          "memory"
        ]
      }
    },
    "/auditor/candidates": {
      "get": {
        "description": "Dry-run microglial surveillance: score the corpus, return the\nranked risk breakdowns. No LLM, no proposals, no watermark movement.",
        "operationId": "auditor_candidates_auditor_candidates_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Auditor Candidates",
        "tags": [
          "memory"
        ]
      }
    },
    "/auditor/run": {
      "post": {
        "description": "Run one auditor pass. mode=auto lets the evidence clock decide;\nlight/deep force the pass (deep = admin request per the cadence spec).",
        "operationId": "auditor_run_auditor_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "auto",
              "pattern": "^(auto|light|deep)$",
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_candidates",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 200,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Candidates"
            }
          },
          {
            "in": "query",
            "name": "max_critic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 50,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Critic"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Auditor Run",
        "tags": [
          "memory"
        ]
      }
    },
...
    "/distill/run": {
      "post": {
        "description": "Distill up to `limit` ready session logs into candidate lessons.",
        "operationId": "distill_run_distill_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 3,
              "maximum": 10,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "min_quiet_minutes",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 1440,
              "minimum": 0,
              "title": "Min Quiet Minutes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Distill Run",
        "tags": [
          "memory"
        ]
      }
    },
    "/distill/status": {
      "get": {
        "description": "Ready-to-distill episode logs (no LLM, no side effects).",
        "operationId": "distill_status_distill_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "min_quiet_minutes",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 1440,
              "minimum": 0,
              "title": "Min Quiet Minutes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Distill Status",
        "tags": [
          "memory"
        ]
      }
    },
...
  }
}